consolidate patches extending the ruleset
