+++ SolastaCommunityExpansion/Patches/CustomFeatures/RulesetCharacterPatcher.cs
@@ -2,11 +2,30 @@
 using HarmonyLib;
 using SolastaCommunityExpansion.CustomFeatureDefinitions;
 
-namespace SolastaCommunityExpansion.Patches.StartOfTurnRecharge
+namespace SolastaCommunityExpansion.Patches.CustomFeatures
 {
     //
     // this patch shouldn't be protected
     //
+    [HarmonyPatch(typeof(RulesetCharacter), "Kill")]
+    [SuppressMessage("Minor Code Smell", "S101:Types should be named in PascalCase", Justification = "Patch")]
+    internal static class RulesetCharacter_Kill
+    {
+        internal static void Prefix(RulesetCharacter __instance)
+        {
+            foreach (var keyValuePair in __instance.ConditionsByCategory)
+            {
+                foreach (RulesetCondition rulesetCondition in keyValuePair.Value)
+                {
+                    if (rulesetCondition?.ConditionDefinition is INotifyConditionRemoval notifiedDefinition)
+                    {
+                        notifiedDefinition.BeforeDyingWithCondition(__instance, rulesetCondition);
+                    }
+                }
+            }
+        }
+    }
+
     [HarmonyPatch(typeof(RulesetCharacter), "RechargePowersForTurnStart")]
     [SuppressMessage("Minor Code Smell", "S101:Types should be named in PascalCase", Justification = "Patch")]
     internal static class RulesetCharacter_RechargePowersForTurnStart
